image-similarity-search/requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
duckdb==1.1.3
filelock==3.16.1
fsspec==2024.12.0
huggingface-hub==0.27.0
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pillow==11.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
ruff==0.8.4
safetensors==0.4.5
sympy==1.13.1
tokenizers==0.21.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.47.1
triton==3.1.0
typing_extensions==4.12.2
urllib3==2.3.0
